example(movies): add more routes for /movie and /tv
