+++ examples/movies/src/components/Hero.module.scss
